change calculate pickup fee

+++ app/Actions/Pricing/PricingCalculator.php
@@ -80,10 +80,10 @@ class PricingCalculator
 
     public static function getPackageTotalAmount(Package $package, bool $is_approved = false)
     {
-        if (! $package->relationLoaded('items.prices')) {
+        if (!$package->relationLoaded('items.prices')) {
             $package->load('items.prices');
         }
-        if (! $package->relationLoaded('prices')) {
+        if (!$package->relationLoaded('prices')) {
             $package->load('prices');
         }
         // get handling and insurance prices
@@ -151,39 +151,36 @@ class PricingCalculator
             'items.*.qty' => ['required', 'numeric'],
             'items.*.handling' => ['nullable']
         ]);
-
+        
         /** @var Price $price */
         $price = self::getPrice($inputs['origin_province_id'], $inputs['origin_regency_id'], $inputs['destination_id']);
-
         $totalWeightBorne = self::getTotalWeightBorne($inputs['items']);
         $insurancePriceTotal = 0;
         $pickup_price = 0;
         if (array_key_exists('fleet_name', $inputs) && $inputs['partner_code'] != '' && $inputs['partner_code'] != null) {
             $partner = Partner::where('code', $inputs['partner_code'])->first();
-            $origin = $inputs['sender_latitude'].', '.$inputs['sender_longitude'];
-            $destination = $partner->latitude.', '.$partner->longitude;
+            $origin = $inputs['sender_latitude'] . ', ' . $inputs['sender_longitude'];
+            $destination = $partner->latitude . ', ' . $partner->longitude;
             $distance = self::distance_matrix($origin, $destination);
-
             if ($inputs['fleet_name'] == 'bike') {
-                if ($distance < 5){
+                if ($distance < 5) {
                     $pickup_price = 8000;
-                }else{
+                } else {
                     $pickup_price = 8000 + (2000 * $distance);
                 }
             } else {
-                if ($distance < 5){
+                if ($distance < 5) {
                     $pickup_price = 15000;
-                }else{
+                } else {
                     $pickup_price = 15000 + (4000 * $distance);
                 }
             }
         }
-
         $discount = 0;
         $handling_price = 0;
 
         foreach ($inputs['items'] as $index => $item) {
-            if (! Arr::has($item, 'handling')) {
+            if (!Arr::has($item, 'handling')) {
                 $item['handling'] = [];
             }
             $handlingResult = [];
@@ -245,9 +242,9 @@ class PricingCalculator
     public static function getServicePrice(array $inputs, ?Price $price = null)
     {
         $inputs =  Validator::validate($inputs, [
-            'origin_province_id' => [Rule::requiredIf(! $price), 'exists:geo_provinces,id'],
-            'origin_regency_id' => [Rule::requiredIf(! $price), 'exists:geo_regencies,id'],
-            'destination_id' => [Rule::requiredIf(! $price), 'exists:geo_sub_districts,id'],
+            'origin_province_id' => [Rule::requiredIf(!$price), 'exists:geo_provinces,id'],
+            'origin_regency_id' => [Rule::requiredIf(!$price), 'exists:geo_regencies,id'],
+            'destination_id' => [Rule::requiredIf(!$price), 'exists:geo_sub_districts,id'],
             'items' => ['required'],
             'items.*.height' => ['required', 'numeric'],
             'items.*.length' => ['required', 'numeric'],
@@ -257,7 +254,7 @@ class PricingCalculator
             'items.*.handling' => ['nullable']
         ]);
 
-        if (! $price) {
+        if (!$price) {
             /** @var Price $price */
             $price = self::getPrice($inputs['origin_province_id'], $inputs['origin_regency_id'], $inputs['destination_id']);
         }
@@ -277,7 +274,7 @@ class PricingCalculator
                 'length' => $item['length'],
                 'width' => $item['width'],
                 'qty' => $item['qty'],
-                'handling' => ! empty($packing) ? array_column($packing, 'type') : null
+                'handling' => !empty($packing) ? array_column($packing, 'type') : null
 
             ];
         }
@@ -304,10 +301,10 @@ class PricingCalculator
         $totalWeightBorne = 0;
 
         foreach ($items as  $item) {
-            if (! Arr::has($item, 'handling')) {
+            if (!Arr::has($item, 'handling')) {
                 $item['handling'] = [];
             }
-            if (! empty($item['handling'])) {
+            if (!empty($item['handling'])) {
                 $item['handling'] = self::checkHandling($item['handling']);
             }
             $totalWeightBorne += self::getWeightBorne($item['height'], $item['length'], $item['width'], $item['weight'], $item['qty'], $item['handling']);
@@ -576,17 +573,18 @@ class PricingCalculator
         return $handling;
     }
 
-    public static function distance_matrix($origin, $destination){
+    public static function distance_matrix($origin, $destination)
+    {
         $response = Http::withHeaders([
             'Accept' => 'application/json',
             'Content-Type' => 'application/json'
-        ])->get('https://maps.googleapis.com/maps/api/distancematrix/json?destinations='.$destination.'&origins='.$origin.'&units=metric&key=AIzaSyAo47e4Aymv12UNMv8uRfgmzjGx75J1GVs');
+        ])->get('https://maps.googleapis.com/maps/api/distancematrix/json?destinations=' . $destination . '&origins=' . $origin . '&units=metric&key=AIzaSyAo47e4Aymv12UNMv8uRfgmzjGx75J1GVs');
         $response = json_decode($response->body());
         $distance = $response->rows[0]->elements[0]->distance->text;
 
-        $distance= str_replace("km","",$distance);
-        $distance= str_replace(",","",$distance);
-        $distance = (double) $distance;
+        $distance = str_replace("km", "", $distance);
+        $distance = str_replace(",", "", $distance);
+        $distance = (float) $distance;
         return $distance;
     }
 }
